Handle scheduler logging in services
